Adding command classes for user, message,  and member commands
